[RFR] Add options prop to TabbedShowLayout (#2740)
* Add options prop to TabbedShowLayout

* Use tabs prop to pass a material-ui's Tabs element

Instead of using an "options" prop to pass props to the <Tabs> in
<TabbedShowLayout>, now a whole <Tabs> custom component can be passed
through the new "tabs" prop.

* Add options prop to TabbedShowLayout

* Use tabs prop to pass a material-ui's Tabs element

Instead of using an "options" prop to pass props to the <Tabs> in
<TabbedShowLayout>, now a whole <Tabs> custom component can be passed
through the new "tabs" prop.

* Remove grandchildren mapping from TabbedShowLayout

Also fix linting.

* Remove unused props and fix misused spread operator

* Pass the rest of the TabbedShowLayoutTabs props to its children

* Pass TabbedShowLayoutTabs rest props to Tabs

Previously it was being passed to its children.

* Fix propTypes
